+++ rest/api/v2010/accounts_conferences_participants.go
@@ -29,7 +29,7 @@ type CreateParticipantParams struct {
 	PathAccountSid *string `json:"PathAccountSid,omitempty"`
 	// The phone number, Client identifier, or username portion of SIP address that made this call. Phone numbers are in [E.164](https://www.twilio.com/docs/glossary/what-e164) format (e.g., +16175551212). Client identifiers are formatted `client:name`. If using a phone number, it must be a Twilio number or a Verified [outgoing caller id](https://www.twilio.com/docs/voice/api/outgoing-caller-ids) for your account. If the `to` parameter is a phone number, `from` must also be a phone number. If `to` is sip address, this value of `from` should be a username portion to be used to populate the P-Asserted-Identity header that is passed to the SIP endpoint.
 	From *string `json:"From,omitempty"`
-	// The phone number, SIP address, or Client identifier that received this call. Phone numbers are in [E.164](https://www.twilio.com/docs/glossary/what-e164) format (e.g., +16175551212). SIP addresses are formatted as `sip:name@company.com`. Client identifiers are formatted `client:name`. [Custom parameters](https://www.twilio.com/docs/voice/api/conference-participant-resource#custom-parameters) may also be specified.
+	// The phone number, SIP address, Client, TwiML App identifier that received this call. Phone numbers are in [E.164](https://www.twilio.com/docs/glossary/what-e164) format (e.g., +16175551212). SIP addresses are formatted as `sip:name@company.com`. Client identifiers are formatted `client:name`. TwiML App identifiers are formatted `app:<APP_SID>`. [Custom parameters](https://www.twilio.com/docs/voice/api/conference-participant-resource#custom-parameters) may also be specified.
 	To *string `json:"To,omitempty"`
 	// The URL we should call using the `status_callback_method` to send status information to your application.
 	StatusCallback *string `json:"StatusCallback,omitempty"`
@@ -123,6 +123,8 @@ type CreateParticipantParams struct {
 	Trim *string `json:"Trim,omitempty"`
 	// A token string needed to invoke a forwarded call. A call_token is generated when an incoming call is received on a Twilio number. Pass an incoming call's call_token value to a forwarded call via the call_token parameter when creating a new call. A forwarded call should bear the same CallerID of the original incoming call.
 	CallToken *string `json:"CallToken,omitempty"`
+	// The name that appears to the called party for this call. Must be between 2 and 255 characters.
+	CallerDisplayName *string `json:"CallerDisplayName,omitempty"`
 }
 
 func (params *CreateParticipantParams) SetPathAccountSid(PathAccountSid string) *CreateParticipantParams {
@@ -321,6 +323,10 @@ func (params *CreateParticipantParams) SetCallToken(CallToken string) *CreatePar
 	params.CallToken = &CallToken
 	return params
 }
+func (params *CreateParticipantParams) SetCallerDisplayName(CallerDisplayName string) *CreateParticipantParams {
+	params.CallerDisplayName = &CallerDisplayName
+	return params
+}
 
 //
 func (c *ApiService) CreateParticipant(ConferenceSid string, params *CreateParticipantParams) (*ApiV2010Participant, error) {
@@ -488,6 +494,9 @@ func (c *ApiService) CreateParticipant(ConferenceSid string, params *CreateParti
 	}
 	if params != nil && params.CallToken != nil {
 		data.Set("CallToken", *params.CallToken)
+	}
+	if params != nil && params.CallerDisplayName != nil {
+		data.Set("CallerDisplayName", *params.CallerDisplayName)
 	}
 
 	resp, err := c.requestHandler.Post(c.baseURL+path, data, headers)