+++ rest/iam/v1/model_iam_v1_create_account_request.go
@@ -17,7 +17,9 @@ package openapi
 // IamV1CreateAccountRequest struct for IamV1CreateAccountRequest
 type IamV1CreateAccountRequest struct {
 	// The owner of the account to be provisioned.
-	OwnerId      string            `json:"ownerId,omitempty"`
-	FriendlyName string            `json:"friendlyName"`
-	Tags         map[string]string `json:"tags,omitempty"`
+	OwnerId string `json:"ownerId,omitempty"`
+	// The parent account ID of the subaccount to be provisioned.
+	ParentAccountId string            `json:"parentAccountId,omitempty"`
+	FriendlyName    string            `json:"friendlyName"`
+	Tags            map[string]string `json:"tags,omitempty"`
 }